pyproject.toml
[project]
name = "coursework-1"
version = "0.1.0"
description = ""
authors = [
    {name = "Kirill",email = "kirill.rylskij96@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
lint = [
    "flake8 (>=7.3.0,<8.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "black (>=25.12.0,<26.0.0)"
]
dev = [
    "pandas-stubs (>=2.3.3.251219,<3.0.0.0)",
    "types-requests (>=2.32.4.20250913,<3.0.0.0)"
]

[tool.black]
line-length = 119
exclude = '''
(
    /(
       \.git
  | \.venv
  | __pycache__
  | build
  | dist
    )/
)
'''

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'